Create Prioritizing Patients
The program begins by printing a short introduction. Then, the program prompts for a patient's name, followed by the patient information that will be used to compute the priority score (see below). The program uses this information to compute a score and determine the patient's priority. This information is printed out, along with a brief recommendation based on the priority. The program then prints a thank you message and prompts for another patient name. This process continues until the word quit is typed as the name, at which point summary statistics for the day are printed .