Update and rename Two Sum to It's Data Time!
The program begins by printing a short introduction and prompting the user for an input file containing a dataset (see below for the expected dataset format). It also prompts the user for an output file. The program will then process the dataset and generate a text-based histogram. It will also compute and print some summary statistics of the data: the number of values read, the mean (average), and the mode (most common value). The program prints the histogram and summary statistics to the output file specified by the user. It then repeatedly prompts the user for an input and output file name until the user enters anything other than "yes" (case-insensitive) to "Do you want to create another histogram?".