Create DNA
For this problem, we will consider a String encoding DNA information as a sequence of As, Cs, Gs, or Ts. For example, we could represent a sequence of DNA as the String "ATGCGCACTATGGTAG" or "TAGC". 

Write a method called mostCommonNucleotide that takes a DNA string as a parameter and returns the nucleotide that appears most frequently in that DNA as a string. Here are some examples of how this method may be called and what it should return.