Create Census
Write a complete runnable Java program named Census (including the class definition and main method) that lets a Census worker type in a sequence of ages for each person they collected data from, and then prints out the average of all of the ages inputted. The program should repeatedly ask for an age of the next person until the Census worker types in a negative number, after which it will print the average age of all of the ages previously typed in.