Create Music Playlist
This program will allow users to manage a music playlist that can queue up songs (put them in the playlist of songs to be played next) and maintain a history of all songs queued. You will use a Queue to maintain the playlist of songs and a Stack to track the user's listening history. The program will allow a user to add songs to the music playlist, play songs, print out the song history, clear the song history, and delete from the song history.